Update DHRobot.py
The reach of the robot is influenced by its tool's dimension. If the robot has all 6 Cartesian coordinates available and its possible to arbitrarily orient its end efector, then by neglecting the difference in distance between the reach space and the work space we can establish that the max length of the tool is the extra reach to be added on the 'd' variable.